fix(demo): render ARIA fallback so the locator picker works from failure pages/clusters
The demo's dom-snapshot handler served only the one committed trace render and
returned no-trace for every other case, so "Pick from snapshot" opened empty
from the failure pages and clusters that showcase locator healing (the
getByLabel and strict-mode clusters have no trace). Reuse the app's own
ARIA-snapshot fallback — already seeded on every failed case — so trace-less
failures render pickable role chips, exactly as production does.

- Extract renderAriaSnapshotHtml into a node-free module (dom-snapshot-aria.ts)
  so the browser demo can import it without pulling in the node-only trace
  parser (zlib); dom-snapshot.ts consumes it internally.
- apiGetDemoDomSnapshot now mirrors resolveCaseDomSnapshot: trace-first, then
  the ARIA fallback from testRunsCases.ariaSnapshot.
- Add unit tests over the exact seeded ARIA shapes.
